Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int x,y,z;
-    scanf("%d %d %d",&x,&y,&z);
+    float x,y,z;
+    scanf("%f %f %f",&x,&y,&z);
     printf("Average: %.2f",(x+y+z)/3);
     return 0;
 }
